Compile assets on every deploy to heroku
To get this working we need to set heroku to use multiple build packs by
following this guide: https://devcenter.heroku.com/changelog-items/653

The commands to run are:

    heroku buildpacks:set https://github.com/heroku/heroku-buildpack-ruby.git
    heroku buildpacks:add --index 1 https://github.com/heroku/heroku-buildpack-nodejs.git

We tested this on a new heroku app, and have applied this change to the
'production' app too.

We also needed to move the node.js dependencies from "devDependencies"
to "dependencies" for the node.js buildpack to pick them up and install
everything. This makes sense as devDependencies is saying that "these
modules are only needed on development/test" and not in production. But
we want gulp et al for production to compile assets.