Fix the failing roster feature spec
These failed because of the new switch logic that keeps the out of hours
rota from being shown inside of standard support hours. I've wrapped it
up in a 'out of hours' context block with a call to Timecop to try to
make it somewhat clearer.  However, I feel that all of this could
be more declarative.  It isn't bad, it just isn't immediately
obvious to someone wihtout IRAT experience why pagerduty would be
preferred over the spreadsheets after hours.