Make the requirements for the dashboards clear
In previous iterations, we seemed to be working to specs built after
prototyping–i.e. they tested everything that was on the page, rather
than the information *sufficient* for the display of the dashboard. This
is a minium specification for what should be shown on the dashboard at
any given time. It does not contain the test logic at this point, but is
only here to document the requirements