Remove default formater and make fuubar avaiable
Progress is the default rspec formatter, so there is no need to declare
it such in the local .rspec.  Doing so overrides other people's choice
of formatter in their global .rspec: fuubar, for example.  I added
fuubar as it is a progress-bar based formatter that reports failures
as they occur and it is my preferred formatter.