Stop it complaining about float api requests
I DRY'd out the api request blocks into a single set of shared examples
and added the requests to float.  A few of them are going to go soon
(like the Google Sheets requests).  For now, though, it seemed sensible
to get all of the respec request logic in one place. That said, the lib
defining the request methods did not seem like the right location.