Remove the unused v2 dashboard code
There was quite a bit of V2 dashboard code that was not acutally being
used anywhere.  I decided to remove it because it was unlikely to ever
be deployed in its current state and it was intertwined in places with
production code (in modes/duty_roster_members.rb, for example), making
them signifcantly longer and more difficult to unpick.