Make TZs consistent, show zone and time to refresh
Working hours were in Zulu time, while out of hours time were not.  This
was causing the live server not to switch at the correct time. I removed
the 'Z' marker from working hours, set zone on the start and end
time comparison methods  I also added information about the current time
and the refresh interval to the dashboard.